+++ src/main/java/net/silentchaos512/gems/lib/soul/GearSoulPart.java
@@ -102,14 +102,6 @@ public class GearSoulPart extends AbstractGearPart {
     }
 
     @Override
-    public void onGearDamaged(PartData part, ItemStack gear, int amount) {
-        SilentGems.LOGGER.debug("Soul onGearDamaged");
-        // Unfortunately no way to get the player, is there?
-        // This means no level-up notifications for armor, for gear souls should work now
-        SoulManager.addSoulXp((int) (GearSoul.XP_FACTOR_ARMOR_DAMAGED * amount), gear, null);
-    }
-
-    @Override
     public int getColor(PartData part, ItemStack gear, int layer, int animationFrame) {
         return Color.VALUE_WHITE;
     }